add account-related logics

+++ todoSpring/src/main/java/com/example/todoSpring/Todo_Sqlite.java
@@ -2,6 +2,7 @@ package com.example.todoSpring;
 
 import org.json.simple.JSONObject;
 
+import java.lang.reflect.Member;
 import java.sql.*;
 import java.util.HashMap;
 import java.util.Map;
@@ -11,23 +12,29 @@ import org.springframework.web.server.ResponseStatusException;
 
 public class Todo_Sqlite {
     public static final String DB_URL = "jdbc:sqlite:todo.db";
+    private Member_Sqlite member_sqlite = new Member_Sqlite();
 
     public void insertDB(TodoDto todoDto) {
         try {
             Connection conn = DriverManager.getConnection(DB_URL);
             Statement stmt = conn.createStatement();
             String sql = "CREATE TABLE IF NOT EXISTS todos (pk INTEGER PRIMARY KEY, " +
-                    "title TEXT NOT NULL, detail TEXT NOT NULL, done BOOLEAN NOT NULL)";
+                    "title TEXT NOT NULL, detail TEXT NOT NULL, done BOOLEAN NOT NULL, " +
+                    "account_id INTEGER, FOREIGN KEY (account_id) REFERENCES members(pk))";
             stmt.executeUpdate(sql);
             stmt.close();
 
-            PreparedStatement pstmt = conn.prepareStatement("INSERT INTO todos(title, detail, done) VALUES(?, ?, ?)");
+            int check_member = member_sqlite.checkMember(todoDto.getId(), todoDto.getPassword());
+
+            PreparedStatement pstmt = conn.prepareStatement("INSERT INTO todos(title, detail, done, account_id)" +
+                    " VALUES(?, ?, ?, ?)");
             pstmt.setString(1, todoDto.getTitle());
             pstmt.setString(2, todoDto.getDetail());
             pstmt.setBoolean(3, todoDto.isDone());
+            pstmt.setInt(4, check_member);
             pstmt.executeUpdate();
             pstmt.close();
-
+            conn.close();
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -44,6 +51,8 @@ public class Todo_Sqlite {
                 Map<String, Object> todo = new HashMap<>();
                 int pk = rs.getInt("pk");
                 todo.put("pk", pk);
+                int account_id = rs.getInt("account_id");
+                todo.put("account_id", account_id);
                 String title = rs.getString("title");
                 todo.put("title", title);
                 String detail = rs.getString("detail");
@@ -62,25 +71,38 @@ public class Todo_Sqlite {
         return new JSONObject(map);
     }
 
-    public JSONObject selectOneDB(int pk) {
-        Map<String, Object> todo = new HashMap<>();
+    public JSONObject selectOneDB(TodoDto todoDto) {
+        Map<String, Object> map = new HashMap<>();
 
         try {
             Connection conn = DriverManager.getConnection(DB_URL);
             Statement stmt = conn.createStatement();
-            ResultSet rs = stmt.executeQuery("SELECT * FROM todos WHERE pk = " + pk + ";");
 
-            if (!rs.next()) {
+            int check_member = member_sqlite.checkMember(todoDto.getId(), todoDto.getPassword());
+
+            ResultSet rs = stmt.executeQuery("SELECT * FROM todos WHERE account_id = " + check_member + ";");
+
+            int result = rs.getInt(1);
+            if(result == 0){
                 throw new ResponseStatusException(HttpStatus.NOT_FOUND);
             }
 
-            todo.put("pk", pk);
-            String title = rs.getString("title");
-            todo.put("title", title);
-            String detail = rs.getString("detail");
-            todo.put("detail", detail);
-            Boolean done = rs.getBoolean("done");
-            todo.put("done", done);
+            while (rs.next()) {
+                Map<String, Object> todo = new HashMap<>();
+                int pk = rs.getInt("pk");
+                todo.put("pk", pk);
+                todo.put("account_id", check_member);
+
+                String title = rs.getString("title");
+                todo.put("title", title);
+
+                String detail = rs.getString("detail");
+                todo.put("detail", detail);
+
+                Boolean done = rs.getBoolean("done");
+                todo.put("done", done);
+                map.put("todo" + pk, todo);
+            }
 
             rs.close();
             stmt.close();
@@ -88,7 +110,7 @@ public class Todo_Sqlite {
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
-        return new JSONObject(todo);
+        return new JSONObject(map);
     }
 
     public void updateDB(int pk, TodoDto todoDto) {
@@ -96,6 +118,25 @@ public class Todo_Sqlite {
             Connection conn = DriverManager.getConnection(DB_URL);
             Statement stmt = conn.createStatement();
 
+            int check_member = member_sqlite.checkMember(todoDto.getId(), todoDto.getPassword());
+
+            String checkSql = "SELECT account_id FROM todos WHERE pk = " + pk + ";";
+            ResultSet rs = stmt.executeQuery(checkSql);
+            int check_account_id = rs.getInt(1);    // 해당 할 일에 대한 account_id 반환
+
+            rs.close();
+            stmt.close();
+
+            // 할 일이 존재하지 않으면 404에러 발생
+            if(check_account_id == 0){
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            }
+            // 계정에 해당 할 일이 없는 경우 403에러 발생
+            if(check_member != check_account_id){
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            }
+
+            Statement stmt2 = conn.createStatement();
             String sql = "UPDATE todos SET ";
             if (todoDto.getTitle() != null){
                 sql +=  "title = '" + todoDto.getTitle() + "',";
@@ -106,35 +147,45 @@ public class Todo_Sqlite {
             sql = sql.substring(0, sql.length() - 1);
             sql += " WHERE pk = " + pk + ";";
 
-            int affectedRows = stmt.executeUpdate(sql);
-
-            stmt.executeUpdate(sql);
-            stmt.close();
+            stmt2.executeUpdate(sql);
+            stmt2.close();
             conn.close();
 
-            if (affectedRows == 0) {
-                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
     }
 
-    public void deleteDB(int pk) {
+    public void deleteDB(int pk, TodoDto todoDto) {
         try {
             Connection conn = DriverManager.getConnection(DB_URL);
             Statement stmt = conn.createStatement();
-            String sql = "DELETE FROM todos WHERE pk = " + pk + ";";
-            int affectedRows = stmt.executeUpdate(sql);
 
-            stmt.executeUpdate(sql);
+            int check_member = member_sqlite.checkMember(todoDto.getId(), todoDto.getPassword());
 
+            String checkSql = "SELECT account_id FROM todos WHERE pk = " + pk + ";";
+            ResultSet rs = stmt.executeQuery(checkSql);
+            int check_account_id = rs.getInt(1);    // 해당 할 일에 대한 account_id 반환
+
+            rs.close();
             stmt.close();
-            conn.close();
 
-            if (affectedRows == 0) {
+            // 할 일이 존재하지 않으면 404에러 발생
+            if(check_account_id == 0){
                 throw new ResponseStatusException(HttpStatus.NOT_FOUND);
             }
+            // 계정에 해당 할 일이 없는 경우 403에러 발생
+            if(check_member != check_account_id){
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            }
+
+            Statement stmt2 = conn.createStatement();
+            String sql = "DELETE FROM todos WHERE pk = " + pk + ";";
+
+            stmt2.executeUpdate(sql);
+            stmt2.close();
+            conn.close();
+
         } catch (SQLException e) {
             e.printStackTrace();
         }
